Cleaned up code-base :wastebasket:
* All tests passed locally :fire:
