+++ .github/.gitignore
@@ -1,1 +1,2 @@
 *.html
+*.Rproj
