+++ man/accuracy.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R, R/S3_accuracy.R
+% Please edit documentation in R/RcppExports.R, R/S3_Accuracy.R
 \name{accuracy.factor}
 \alias{accuracy.factor}
 \alias{weighted.accuracy.factor}
