+++ man/precision.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R, R/S3_precision.R
+% Please edit documentation in R/RcppExports.R, R/S3_Precision.R
 \name{precision.factor}
 \alias{precision.factor}
 \alias{weighted.precision.factor}
