+++ man/recall.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R, R/S3_recall.R
+% Please edit documentation in R/RcppExports.R, R/S3_Recall.R
 \name{recall.factor}
 \alias{recall.factor}
 \alias{weighted.recall.factor}
