+++ man/specificity.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R, R/S3_specificity.R
+% Please edit documentation in R/RcppExports.R, R/S3_Specificity.R
 \name{specificity.factor}
 \alias{specificity.factor}
 \alias{weighted.specificity.factor}
